+++ docs/misc/changelog.rst
@@ -12,10 +12,12 @@ Breaking Changes:
 
 New Features:
 ^^^^^^^^^^^^^
+- Allow custom actor/critic network architectures using ``net_arch=dict(qf=[400, 300], pi=[64, 64])`` for off-policy algorithms (SAC, TD3, DDPG)
 - Added Hindsight Experience Replay ``HER``. (@megan-klaiber)
 
 Bug Fixes:
 ^^^^^^^^^^
+- Fix GAE computation for on-policy algorithms (off-by one for the last value) (thanks @Wovchena)
 
 Deprecations:
 ^^^^^^^^^^^^^